% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelPred_accessors.R
\docType{methods}
\name{getOutcome}
\alias{getOutcome}
\alias{getOutcome,modelPredictions-method}
\title{Set the generic for accessor (getter) for outcomes}
\usage{
getOutcome(object)

\S4method{getOutcome}{modelPredictions}(object)
}
\description{
Set the generic for accessor (getter) for outcomes
}
\section{Methods (by class)}{
\itemize{
\item \code{modelPredictions}: Define the method for getting predicted outcomes
}}

