% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelPred_accessors.R
\docType{methods}
\name{getProb}
\alias{getProb}
\alias{getProb,modelPredictions-method}
\title{Set the generic for accessor (getter) for probabilities}
\usage{
getProb(object)

\S4method{getProb}{modelPredictions}(object)
}
\description{
Set the generic for accessor (getter) for probabilities
}
\section{Methods (by class)}{
\itemize{
\item \code{modelPredictions}: Define the method for getting predicted probabilities
}}

