% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelPred_accessors.R
\docType{methods}
\name{getStats}
\alias{getStats}
\alias{getStats,modelPredictions-method}
\title{Set the generic for accessor (getter) for statistcs (acc, sens, spec, kappa, AUC)}
\usage{
getStats(object)

\S4method{getStats}{modelPredictions}(object)
}
\description{
Set the generic for accessor (getter) for statistcs (acc, sens, spec, kappa, AUC)
}
\section{Methods (by class)}{
\itemize{
\item \code{modelPredictions}: Define the method for getting test set statistics
}}

