% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelPred_class.R
\docType{methods}
\name{initialize,modelPredictions-method}
\alias{initialize,modelPredictions-method}
\title{Initializer for modelPredictions class}
\usage{
\S4method{initialize}{modelPredictions}(.Object, modelName, outcomes,
  trueOutcome, predOutcome, predProb, testStats)
}
\description{
Initializer for modelPredictions class
}

