% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelPred_constructor.R
\name{modelPredConstr}
\alias{modelPredConstr}
\title{Constructor function for objects of class modelPrediction. Always assume first class is class of interest.}
\usage{
modelPredConstr(models, inputData, trueResult, classLevels)
}
\description{
Constructor function for objects of class modelPrediction. Always assume first class is class of interest.
}

