% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelPred_class.R
\docType{class}
\name{modelPredictions-class}
\alias{modelPredictions-class}
\title{S4 class definition for model predictions. Only applicable for two class problems}
\description{
S4 class definition for model predictions. Only applicable for two class problems
}
\section{Slots}{

\describe{
\item{\code{modelName}}{a character vector containing the names of the fitted models}

\item{\code{outcomes}}{a character vector indicating the possible outcomes}

\item{\code{trueOutcome}}{a numeric vector indicating the true outcome. Stored as binary 1/0 correspoinding to positive/ negative class}

\item{\code{predOutcome}}{a matrix corresponding to the predictions (binary 1/0) from each model}

\item{\code{predProb}}{a martix storing the predicted probabilities of the positive class}

\item{\code{testStats}}{a matrix storing the sens, spec, acc, auc and kappa for each model's predictions}
}}

