% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelPred_generics.R
\docType{methods}
\name{plot,modelPredictions,ANY-method}
\alias{plot,modelPredictions,ANY-method}
\title{Plot method for objects of class modelPredictions}
\usage{
\S4method{plot}{modelPredictions,ANY}(x, y, ...)
}
\description{
Plot method for objects of class modelPredictions
}

