% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/modelPred_generics.R
\docType{methods}
\name{show,modelPredictions-method}
\alias{show,modelPredictions-method}
\title{Show method for modelPredictions}
\usage{
\S4method{show}{modelPredictions}(object)
}
\description{
Show method for modelPredictions
}

